src/requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.8
django-import-export==1.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
et-xmlfile==1.0.1
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.3
PyJWT==1.7.1
pytz==2019.2
PyYAML==5.1.2
tablib==0.13.0
UNKNOWN==0.0.0
xlrd==1.2.0
xlwt==1.3.0
